Updated formatting and style of various pages. Added new options to web map and fixed various issues.
Updated the formatting and style used by the about, contact, references pages. Added new info to about page. Fixed the manual buffer option to not appear in the legend when the value was zero. Added the satellite option to allow users to choose between satellite basemap and the default. Changed the default basemap to the Mapbox monochrome dark theme to better fit with the other pages and increase visibility of features within the map. Added instruction popup window to the web map. Added additional interactable help icons to the web map options panel.